+++ pkgs/ffigen/lib/src/visitor/copy_methods_from_super_type.dart
@@ -49,18 +49,18 @@ class CopyMethodsFromSuperTypesVisitation extends Visitation {
     if (superType != null) {
       for (final m in superType.methods) {
         if (isNSObject) {
-          node.addMethod(m);
+          node.copyMethod(m);
         } else if (m.isClassMethod &&
             !_excludedNSObjectMethods.contains(m.originalName)) {
-          node.addMethod(m);
+          node.copyMethod(m);
         } else if (ObjCBuiltInFunctions.isInstanceType(m.returnType)) {
-          node.addMethod(m);
+          node.copyMethod(m);
         }
       }
     }
 
     // Copy all methods from all the interface's protocols.
-    _copyMethodFromProtocols(node, node.protocols, node.addMethod);
+    _copyMethodFromProtocols(node, node.protocols, node.copyMethod);
 
     // Copy methods from all the categories that extend this interface, if those
     // methods return instancetype, because the Dart inheritance rules don't
@@ -74,7 +74,7 @@ class CopyMethodsFromSuperTypesVisitation extends Visitation {
     for (final category in node.categories) {
       for (final m in category.methods) {
         if (category.shouldCopyMethodToInterface(m)) {
-          node.addMethod(m);
+          node.copyMethod(m);
         }
       }
     }
@@ -83,16 +83,16 @@ class CopyMethodsFromSuperTypesVisitation extends Visitation {
   void _copyMethodFromProtocols(
     Binding node,
     List<ObjCProtocol> protocols,
-    void Function(ObjCMethod) addMethod,
+    void Function(ObjCMethod) copyMethod,
   ) {
     // Copy all methods from all the protocols.
     final isNSObject = ObjCBuiltInFunctions.isNSObject(node.originalName);
     for (final proto in protocols) {
       for (final m in proto.methods) {
         if (isNSObject) {
-          addMethod(m);
+          copyMethod(m);
         } else if (!_excludedNSObjectMethods.contains(m.originalName)) {
-          addMethod(m);
+          copyMethod(m);
         }
       }
     }
@@ -103,7 +103,7 @@ class CopyMethodsFromSuperTypesVisitation extends Visitation {
     node.visitChildren(visitor);
 
     // Copy all methods from all the category's protocols.
-    _copyMethodFromProtocols(node, node.protocols, node.addMethod);
+    _copyMethodFromProtocols(node, node.protocols, node.copyMethod);
   }
 
   @override
@@ -124,7 +124,7 @@ class CopyMethodsFromSuperTypesVisitation extends Visitation {
       // So copy across all the methods explicitly, rather than trying to use
       // Dart inheritance to get them implicitly.
       for (final method in superProtocol.methods) {
-        node.addMethod(method);
+        node.copyMethod(method);
       }
     }
   }